Change form field validation strategy.
During an onBlur a field is validated, and set to invalid if validation
failed. During an onChange the filed is only validated, if it already
invalid. That means, when an onBlur changes a field to invalid and the
fixes the field content, then each changes triggers a validation to
show the user immediately  when he fixed th field.