Token refresh handling during call of REST endpoints from middleware.
Middleware calls an REST endpoint if the provided action is a
'RestRequestAction' of the type 'REST_REQUEST'. If so, the action
contains all necessary information to call the REST endpoint.

Also token handling is done. If the 'REST_REQUEST' needs a BEARER
authentication, the access_token is fetched from the store and will be
provided as Authentication header.

For the case that the access_token is invalid, so an 401 is returned by
the server, we try to get a new access_token with the refresh_token
laying in the store. If the server send a new access_token we update the
token on the store and redo the current request.

If the server answers also answers with an 401 during token refresh
(e.g. the refresh token is invalid), the user is immediately logged out.

Finally when the server answers with an 403, that means someone tries an
unauthorized access, the user is also immediately logged out.